rentme/backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
supabase==2.3.4
python-dotenv==1.0.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI
openai==1.12.0

# Utils
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
